Refactor ContestCreation
