+++ contestmanagementsystem/models/Association.ts
@@ -1,11 +1,11 @@
-// models/Organization.ts
+// models/Association.ts
 import mongoose from 'mongoose';
 
-const OrganizationSchema = new mongoose.Schema({
+const AssociationSchema = new mongoose.Schema({
   name: { type: String, required: true },
   description: String,
   website: String,
   logo: String, // Will be updated when upload is handled
 }, { timestamps: true });
 
-export default mongoose.models.Organization || mongoose.model('Organization', OrganizationSchema);
+export const Association = mongoose.models.Association || mongoose.model('Association', AssociationSchema);
